% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrices.R
\name{accuracy}
\alias{accuracy}
\title{Get Accuracy}
\usage{
accuracy(conf_mat)
}
\arguments{
\item{conf_mat}{A confusion matrix}
}
\value{
The accuracy
}
\description{
Calculate accuracy
}
\author{
Stefanie Molin
}
