% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_study_results.R
\name{arrange_AS_alert_count_graphs}
\alias{arrange_AS_alert_count_graphs}
\title{AS Alert Counts Graphs Arranger}
\usage{
arrange_AS_alert_count_graphs(overall_results_count, metrics_count)
}
\arguments{
\item{overall_results_count}{List of results from AS_alert_counts()}

\item{metrics_count}{List with results from AS_alert_counts_by_metric()}
}
\value{
A plot object
}
\description{
Places the 4 graphs for the alert counts and distributions in 1 object
}
\author{
Stefanie Molin
}
