% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_library.R
\name{arrange_AS_violin_plots}
\alias{arrange_AS_violin_plots}
\title{AS Violin Plots Arrangement}
\usage{
arrange_AS_violin_plots(data)
}
\arguments{
\item{data}{AS dataframe}
}
\value{
A plot object
}
\description{
Creates arrangement of violin plots of AS data
}
\author{
Stefanie Molin
}
