% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_library.R
\name{arrange_RexT_violin_plots}
\alias{arrange_RexT_violin_plots}
\title{RexT Violin Plots Arrangement}
\usage{
arrange_RexT_violin_plots(data)
}
\arguments{
\item{data}{RexT dataframe}
}
\value{
A plot object
}
\description{
Creates arrangement of violin plots of RexT data
}
\author{
Stefanie Molin
}
