% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_library.R
\name{arrange_TS_violin_plots}
\alias{arrange_TS_violin_plots}
\title{TS Violin Plots Arrangement}
\usage{
arrange_TS_violin_plots(data)
}
\arguments{
\item{data}{TS dataframe}
}
\value{
A plot object
}
\description{
Creates arrangement of violin plots of TS data
}
\author{
Stefanie Molin
}
