% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_study_results.R
\name{AS_alert_counts_by_metric}
\alias{AS_alert_counts_by_metric}
\title{AS Metric Alert Counts}
\usage{
AS_alert_counts_by_metric(beginning, middle, end)
}
\arguments{
\item{beginning}{Dataframe of AS results for beginning of the month}

\item{middle}{Dataframe of AS results for middle of the month}

\item{end}{Dataframe of AS results for end of the month}
}
\value{
A list containing dataframe of client alert counts, a dataframe of campaign alert counts, and a graph of the results
}
\description{
Calculate AS alert counts by metric for case study.
}
\author{
Stefanie Molin
}
