% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrices.R
\name{AS_overall_confusion_matrix}
\alias{AS_overall_confusion_matrix}
\title{AS Period Confusion Matrix}
\usage{
AS_overall_confusion_matrix(beginning, middle, end, logs)
}
\arguments{
\item{beginning}{Dataframe of AS results for beginning of the month}

\item{middle}{Dataframe of AS results for middle of the month}

\item{end}{Dataframe of AS results for end of the month}

\item{logs}{Dataframe of logs from Metis}
}
\value{
A dataframe
}
\description{
Calculate AS confusion matrix by date
}
\author{
Stefanie Molin
}
