% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrices.R
\name{calculate_accuracy}
\alias{calculate_accuracy}
\title{Calculate Accuracy}
\usage{
calculate_accuracy(TP, TN, FP, FN)
}
\arguments{
\item{TP}{Number of true positives}

\item{TN}{Number of true negatives}

\item{FP}{Number of false positives}

\item{FN}{Number of false negatives}
}
\value{
The accuracy
}
\description{
Calculate accuracy
}
\author{
Stefanie Molin
}
