% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_change_models.R
\name{calculate_aggregations}
\alias{calculate_aggregations}
\title{Calculate Aggregations}
\usage{
calculate_aggregations(data, FUN)
}
\arguments{
\item{data}{Data to use for aggregations}

\item{FUN}{Function to use for aggregations}
}
\value{
A dataframe of aggregated data
}
\description{
Calculate aggregations be specified function at 7, 30, and 60 day lookbacks
}
\author{
Stefanie Molin
}
