% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrices.R
\name{calculate_F1_score}
\alias{calculate_F1_score}
\title{Calculate F1-score}
\usage{
calculate_F1_score(precision, recall)
}
\arguments{
\item{precision}{The precision}

\item{recall}{The recall}
}
\value{
The F1-score
}
\description{
Calculate F1-score
}
\author{
Stefanie Molin
}
