% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_library.R
\name{calculate_mode}
\alias{calculate_mode}
\title{Mode}
\usage{
calculate_mode(x)
}
\arguments{
\item{x}{Vector of data to calculate the mode from}
}
\value{
The mode (most common value)
}
\description{
Calculate the mode of a vector
}
\author{
Stefanie Molin
}
