% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrices.R
\name{calculate_precision}
\alias{calculate_precision}
\title{Calculate Precision}
\usage{
calculate_precision(TP, FP)
}
\arguments{
\item{TP}{Number of true positives}

\item{FP}{Number of false positives}
}
\value{
The precision
}
\description{
Calculate precision
}
\author{
Stefanie Molin
}
