% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrices.R
\name{calculate_recall}
\alias{calculate_recall}
\title{Calculate Recall}
\usage{
calculate_recall(TP, FN)
}
\arguments{
\item{TP}{Number of true positives}

\item{FN}{Number of false negatives}
}
\value{
The recall
}
\description{
Calculate recall
}
\author{
Stefanie Molin
}
