% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrices.R
\name{calculate_specificity}
\alias{calculate_specificity}
\title{Calculate Specificity}
\usage{
calculate_specificity(TN, FP)
}
\arguments{
\item{TN}{Number of true negatives}

\item{FP}{Number of false positives}
}
\value{
The specificity
}
\description{
Calculate specificity
}
\author{
Stefanie Molin
}
