% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_study_options.R
\name{case_study_log_level}
\alias{case_study_log_level}
\title{Set Logging Level}
\usage{
case_study_log_level(level_name)
}
\arguments{
\item{level_name}{String such as "INFO" or "DEBUG"}
}
\description{
Set logging level for case study.
}
\author{
Stefanie Molin
}
