% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{case_study_theme}
\alias{case_study_theme}
\title{All Centered Case Study Theme}
\usage{
case_study_theme()
}
\value{
A ggplot theme
}
\description{
Theme for case study graphs with titles, subtitles and captions centered.
}
\author{
Stefanie Molin
}
