% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_data_for_case_study}
\alias{clean_data_for_case_study}
\title{Clean Data}
\usage{
clean_data_for_case_study(df)
}
\arguments{
\item{df}{Dataframe to clean}
}
\value{
A dataframe
}
\description{
Replace Inf, NaN and NA values with 0
}
\author{
Stefanie Molin
}
