% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_library.R
\name{coefficient_of_variation_stats}
\alias{coefficient_of_variation_stats}
\title{Coefficient of Variation}
\usage{
coefficient_of_variation_stats(x)
}
\arguments{
\item{x}{Vector of data}
}
\value{
The coefficient of variation
}
\description{
Calculate the coefficient of variation of a vector
}
\author{
Stefanie Molin
}
