% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_library.R
\name{coefficient_of_variation_violin_plots}
\alias{coefficient_of_variation_violin_plots}
\title{Coefficient of Variation Violin Plots}
\usage{
coefficient_of_variation_violin_plots(data, series_column, metrics,
  last_index = NULL, preprocess = FALSE)
}
\arguments{
\item{data}{Wide dataframe (unmelted)}

\item{series_column}{String name of the column containing the unique identifier of the series}

\item{metrics}{Vector of the metrics that are to be shown}

\item{last_index}{The index of the last metric to show in the table. 
Defaults to NULL meaning you will only be returned one graph. 
If you choose to split, you will be returned a plot object of multiple plots.}

\item{preprocess}{Boolean indicating whether or not to prepocess the data, use this if the data won't work as is. Defaults to FALSE.}
}
\value{
A plot object
}
\description{
Creates violin plots showing the distribution of the coefficient of variation across the given series and metrics
}
\author{
Stefanie Molin
}
