% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_plot.R
\name{control_chart}
\alias{control_chart}
\title{Control Chart}
\usage{
control_chart(data)
}
\arguments{
\item{data}{Dataframe with 2 columns: "index" and "value",
where "index" is the number of the row and "value" is the value to plot}
}
\value{
A ggplot object
}
\description{
Plots a control chart with colored lines for zones.
}
\note{
More info here: \url{https://en.wikipedia.org/wiki/Western_Electric_rules} and here: 
\url{https://en.wikipedia.org/wiki/Nelson_rules}
}
\author{
Stefanie Molin
}
