% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_library.R
\name{correlation_matrix}
\alias{correlation_matrix}
\title{Correlation Matrix}
\usage{
correlation_matrix(data)
}
\arguments{
\item{data}{Correlation data}
}
\value{
A correlation matrix
}
\description{
Correlation of AS, TS, and RexT metrics
}
\author{
Stefanie Molin
}
