% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_study_detect_anomaly.R
\name{detect_anomaly}
\alias{detect_anomaly}
\title{Detect Anomaly}
\usage{
detect_anomaly(data, metric, yesterday)
}
\arguments{
\item{data}{Dataframe where first column is "day" and second is the metric you want to check for anomalies}

\item{metric}{Can be "events", "COS", "CR", "CTR", "spend", "rext", "exec". Which metric to check for anomaly.}

\item{yesterday}{The date to check for alerts.}
}
\value{
A boolean indicating whether or not yesterday was an alert
}
\description{
Determine whether or not yesterday was an anomaly after cleaning data.
}
\author{
Stefanie Molin
}
