% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{disqualify_accounts_for_case_study}
\alias{disqualify_accounts_for_case_study}
\title{Disqualify Accounts for Case Study}
\usage{
disqualify_accounts_for_case_study(data, series_column, is_AS_data = TRUE,
  required_entries = 25)
}
\arguments{
\item{data}{Dataframe to search. Must have columns: tac and day.}

\item{series_column}{Column name of the series to check for disqualification}

\item{is_AS_data}{Boolean for whether or not this is AS data. AS data will have a TAC column for an additional disqualification.}

\item{required_entries}{Number of days the series must have data for in the last 30 days to be considered live}
}
\value{
A dataframe
}
\description{
Return the series that aren't eligible for alerts (TAC of 0 for last 2 days and/or not live for 25 days in the last 30 days)
}
\author{
Stefanie Molin
}
