% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_library.R
\name{EDA_data_prep_AS}
\alias{EDA_data_prep_AS}
\title{EDA Data Prep for AS Data}
\usage{
EDA_data_prep_AS(df, beginning_dates, middle_dates, end_dates)
}
\arguments{
\item{df}{Dataframe of AS data}

\item{beginning_dates}{Dates to marked as "beginning" period}

\item{middle_dates}{Dates to marked as "middle" period}

\item{end_dates}{Dates to marked as "end" period}
}
\value{
A list with client-level and campaign-level dataframes
}
\description{
Prepare the AS data for EDA
}
\author{
Stefanie Molin
}
