% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_library.R
\name{EDA_data_prep_RexT}
\alias{EDA_data_prep_RexT}
\title{EDA Data Prep for RexT Data}
\usage{
EDA_data_prep_RexT(df, beginning_dates, middle_dates, end_dates)
}
\arguments{
\item{df}{Dataframe of RexT data}

\item{beginning_dates}{Dates to marked as "beginning" period}

\item{middle_dates}{Dates to marked as "middle" period}

\item{end_dates}{Dates to marked as "end" period}
}
\value{
A dataframe
}
\description{
Prepare the RexT data for EDA
}
\author{
Stefanie Molin
}
