% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_frequency.R
\name{estimate_frequency}
\alias{estimate_frequency}
\title{Estimate Frequency}
\usage{
estimate_frequency(data)
}
\arguments{
\item{data}{Vector of values in order.}
}
\value{
A dataframe
}
\description{
Estimate the frequency of vector
}
\author{
Stefanie Molin
}
