% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_study_results.R
\name{exec_alert_counts}
\alias{exec_alert_counts}
\title{Exec Alert Counts}
\usage{
exec_alert_counts(data, beginning_dates, middle_dates, end_dates)
}
\arguments{
\item{data}{Dataframe of territory RexT results}

\item{beginning_dates}{Vector of dates for beginning of the month}

\item{middle_dates}{Vector of dates for middle of the month}

\item{end_dates}{Vector of dates for end of the month}
}
\value{
A list containing dataframe of TS alert counts, a graph of the results, and a boxplot of the distribution
}
\description{
Calculate territory RexT alert counts for case study.
}
\author{
Stefanie Molin
}
