% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_study_results.R
\name{exec_alert_counts_by_territory}
\alias{exec_alert_counts_by_territory}
\title{RexT Alerts by Territory}
\usage{
exec_alert_counts_by_territory(data, beginning_dates, middle_dates, end_dates)
}
\arguments{
\item{data}{Dataframe of territory RexT results}

\item{beginning_dates}{Vector of dates for beginning of the month}

\item{middle_dates}{Vector of dates for middle of the month}

\item{end_dates}{Vector of dates for end of the month}
}
\value{
A list of a dataframe containing the alert counts and a graph.
}
\description{
Calculate RexT alert counts by territory for case study.
}
\author{
Stefanie Molin
}
