% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrices.R
\name{exec_overall_confusion_matrix}
\alias{exec_overall_confusion_matrix}
\title{Territory RexT Period Confusion Matrix}
\usage{
exec_overall_confusion_matrix(data, logs, beginning_dates, middle_dates,
  end_dates)
}
\arguments{
\item{data}{Dataframe of territory RexT results}

\item{logs}{Dataframe of logs from Metis}

\item{beginning_dates}{Vector of dates for beginning of the month}

\item{middle_dates}{Vector of dates for middle of the month}

\item{end_dates}{Vector of dates for end of the month}
}
\value{
A dataframe
}
\description{
Calculate Territory RexT confusion matrix by date
}
\author{
Stefanie Molin
}
