% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrices.R
\name{exec_territories_confusion_matrix}
\alias{exec_territories_confusion_matrix}
\title{Territory RexT Metrics Confusion Matrix}
\usage{
exec_territories_confusion_matrix(data, logs)
}
\arguments{
\item{data}{Dataframe of territory RexT results}

\item{logs}{Dataframe of logs from Metis}
}
\value{
A dataframe
}
\description{
Calculate Territory RexT confusion matrix by metric
}
\author{
Stefanie Molin
}
