% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrices.R
\name{f1_score}
\alias{f1_score}
\title{Get F1 Score}
\usage{
f1_score(conf_mat)
}
\arguments{
\item{conf_mat}{A confusion matrix}
}
\value{
The F1 Score metric
}
\description{
Calculate F1 score
}
\author{
Stefanie Molin
}
