% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{facet_wrap_theme}
\alias{facet_wrap_theme}
\title{Facet Wrap Theme}
\usage{
facet_wrap_theme()
}
\value{
A ggplot formatting
}
\description{
Formatting of facet wrap for case study
}
\author{
Stefanie Molin
}
