% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrices.R
\name{false_negatives}
\alias{false_negatives}
\title{Get False Negatives}
\usage{
false_negatives(conf_mat)
}
\arguments{
\item{conf_mat}{A confusion matrix}
}
\value{
The number of false negatives
}
\description{
Calculate false negatives
}
\author{
Stefanie Molin
}
