% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_curves.R
\name{false_positive_rate}
\alias{false_positive_rate}
\title{Get False Positive Rate}
\usage{
false_positive_rate(false_positives, true_negatives)
}
\arguments{
\item{false_positives}{Number of false positives}

\item{true_negatives}{Number of true negatives}
}
\value{
FPR
}
\description{
Calculate false positive rate (FPR)
}
\author{
Stefanie Molin
}
