% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrices.R
\name{false_positives}
\alias{false_positives}
\title{Get False Positives}
\usage{
false_positives(conf_mat)
}
\arguments{
\item{conf_mat}{A confusion matrix}
}
\value{
The number of false positives
}
\description{
Calculate false positives
}
\author{
Stefanie Molin
}
