% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_change_models.R
\name{flag_percent_change_alerts}
\alias{flag_percent_change_alerts}
\title{Flag Percent Change Alerts}
\usage{
flag_percent_change_alerts(data, aggregation, threshold)
}
\arguments{
\item{data}{Dataframe containing the percent change info}

\item{aggregation}{String specifying aggregation level to use. Can be 'L7D', 'L30D', or 'L60D'}

\item{threshold}{Percentage threshold for alerts as decimal.}
}
\value{
The provided dataframe with an 'is_alert' column added
}
\description{
Calculate alerts for percent difference in excess of the given threshold for the chosen lookback window
}
\author{
Stefanie Molin
}
