% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{lookback_window_color_scheme}
\alias{lookback_window_color_scheme}
\title{Color Scheme for Lookback Window ROC Curves}
\usage{
lookback_window_color_scheme()
}
\value{
A ggplot color scheme
}
\description{
Color scheme for qualitative uses
}
\author{
Stefanie Molin
}
