% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_library.R
\name{metrics_correlation_data}
\alias{metrics_correlation_data}
\title{Correlation Calculations}
\usage{
metrics_correlation_data(AS_data, TS_data, RexT_data)
}
\arguments{
\item{AS_data}{Wide dataframe of AS data}

\item{TS_data}{Dataframe of TS data}

\item{RexT_data}{Territory RexT dataframe}
}
\value{
A list of correlation data
}
\description{
Correlations of AS, TS, and RexT metrics
}
\author{
Stefanie Molin
}
