% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrices.R
\name{precision}
\alias{precision}
\title{Get Precision}
\usage{
precision(conf_mat)
}
\arguments{
\item{conf_mat}{A confusion matrix}
}
\value{
The precision metric
}
\description{
Calculate precision
}
\author{
Stefanie Molin
}
