% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_to_python_data_prep.R
\name{prep_RexT_for_python}
\alias{prep_RexT_for_python}
\title{Prep RexT Data for Transfer to Python for Modelling}
\usage{
prep_RexT_for_python(data, days)
}
\arguments{
\item{data}{Dataframe of daily data to check for anomalies.}

\item{days}{Number of days to require for model (others will be marked as disqualified)}
}
\value{
Prepped data as dataframe
}
\description{
Clean data and mark disqualifications for transfer to Python for modelling
}
\note{
This function relies on the structure of the query that provides "data". 
If changing column names, adding or subtracting columns, be sure to update here as well.
}
\author{
Stefanie Molin
}
