% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_curves.R
\name{prep_ROC_data}
\alias{prep_ROC_data}
\title{Prep ROC Data for points graph}
\usage{
prep_ROC_data(data, name, facet_column = NULL)
}
\arguments{
\item{data}{Dataframe to use for ROC calculation}

\item{name}{What to name the series in the graph}

\item{facet_column}{The name of the column to use for the facets}
}
\value{
TPR
}
\description{
Format data for ROC points graph
}
\author{
Stefanie Molin
}
