% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/western_electric_rules.R
\name{previous_western_electric_rules}
\alias{previous_western_electric_rules}
\title{Previous Western Electric Rules}
\usage{
previous_western_electric_rules(rule, rules_order, include = TRUE)
}
\arguments{
\item{rule}{Name of the rule to start with}

\item{rules_order}{Vector of rules from first checked to last}

\item{include}{Boolean indicating whether or not to include the rule provided in the list. Defaults to TRUE (include it)}
}
\value{
A vector of rules or "N/A" if there are no rules and the provided one isn't included
}
\description{
Return the current rules and previous rules that would preceed this one.
}
\author{
Stefanie Molin
}
