% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_metis.R
\name{query_metis}
\alias{query_metis}
\title{Query Metis}
\usage{
query_metis(query)
}
\arguments{
\item{query}{PostgreSQL query as a string}
}
\value{
Dataframe with the results
}
\description{
Query Metis for case study data and classifications
}
\author{
Stefanie Molin
}
