% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrices.R
\name{recall}
\alias{recall}
\title{Get Recall}
\usage{
recall(conf_mat)
}
\arguments{
\item{conf_mat}{A confusion matrix}
}
\value{
The recall metric
}
\description{
Calculate recall
}
\author{
Stefanie Molin
}
