% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_library.R
\name{reshape_TS_data_for_EDA}
\alias{reshape_TS_data_for_EDA}
\title{TS Data Reshape for EDA}
\usage{
reshape_TS_data_for_EDA(data)
}
\arguments{
\item{data}{Wide dataframe (unmelted)}
}
\value{
A dataframe
}
\description{
Reshapes the TS data for EDA purposes only
}
\author{
Stefanie Molin
}
