% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_curves.R
\name{ROC_curve}
\alias{ROC_curve}
\title{Plot ROC Curve}
\usage{
ROC_curve(data, logs, curve_name_column, prediction_column, method_name,
  graph_title = "ROC Curve by Product", thresholds = seq(from = 0, to = 1,
  by = 0.01))
}
\arguments{
\item{data}{Data with column for metric used with threshold to determine alert}

\item{logs}{Log data from Metis.}

\item{curve_name_column}{String name of the column to use as a series (each line of the curve).}

\item{prediction_column}{String name of the column to use as a the prediction.}

\item{method_name}{Name of the method of triggering alerts. Used as subtitle for graph.}

\item{graph_title}{Title of the graph. Defaults to "ROC Curve by Product"}

\item{thresholds}{Vector of thresholds to use for drawing the ROC curve. Defaults to 0-100\%.}
}
\value{
ROC Curve plot
}
\description{
Graph results on ROC curve for given thresholds
}
\author{
Stefanie Molin
}
