% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_curves.R
\name{ROC_curve_points_facet}
\alias{ROC_curve_points_facet}
\title{Plot ROC points Faceted}
\usage{
ROC_curve_points_facet(AS_data, TS_data, exec_data, facet_column, method_name)
}
\arguments{
\item{AS_data}{Comparison results of AS alerts to the logs.}

\item{TS_data}{Comparison results of TS alerts to the logs.}

\item{exec_data}{Comparison results of exec alerts to the logs.}

\item{facet_column}{The name of the column to use for the facets}

\item{method_name}{Name of the method of triggering alerts. Used as subtitle for graph.}
}
\value{
ROC Curve plot
}
\description{
Graph results on ROC as points when thresholds aren't used by the model to determine alerts
}
\author{
Stefanie Molin
}
