% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_study_detect_anomaly.R
\name{run_anomaly_detection}
\alias{run_anomaly_detection}
\title{Run Anomaly Detection}
\usage{
run_anomaly_detection(data, direction, stat_sig, yesterday)
}
\arguments{
\item{data}{Dataframe where first column is "day" and second is the metric you want to check for anomalies}

\item{direction}{Can be "both", "neg", or "pos". Determines whether you only want to check for positive anomalies, negative, or both.}

\item{stat_sig}{The statistical significance for the anomaly test.}

\item{yesterday}{The date to check for alerts.}
}
\value{
A boolean indicating whether or not yesterday was an alert
}
\description{
Determine whether or not yesterday was an anomaly
}
\author{
Stefanie Molin
}
