% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_library.R
\name{seasonality_violin_plots}
\alias{seasonality_violin_plots}
\title{Seasonality Violin Plots}
\usage{
seasonality_violin_plots(AS_data, TS_data, RexT_data)
}
\arguments{
\item{AS_data}{Long dataframe of AS data}

\item{TS_data}{Dataframe of TS data}

\item{RexT_data}{Territory RexT dataframe}
}
\value{
A plot object
}
\description{
Creates violin plots showing the distribution of the metrics by day of week
}
\author{
Stefanie Molin
}
