% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrices.R
\name{specificity}
\alias{specificity}
\title{Get Specificity}
\usage{
specificity(conf_mat)
}
\arguments{
\item{conf_mat}{A confusion matrix}
}
\value{
The specificity
}
\description{
Calculate specificity
}
\author{
Stefanie Molin
}
