% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_normal_plot.R
\name{standard_normal_plot}
\alias{standard_normal_plot}
\title{Standard Normal Plot}
\usage{
standard_normal_plot()
}
\value{
A ggplot object
}
\description{
Plots a standard normal with colored sections for standard deviations
}
\author{
Stefanie Molin
}
