% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{time_of_month_color_scheme}
\alias{time_of_month_color_scheme}
\title{Color Scheme for Time of Month}
\usage{
time_of_month_color_scheme(reverse = FALSE)
}
\arguments{
\item{reverse}{Boolean indicating whether or not to reverse the color order for the chart (useful when some of the charts will use coord_flip() but the legend will be shared).
Defaults to FALSE}
}
\value{
A ggplot color scheme
}
\description{
Color scheme for beginning, middle, and end of month
}
\author{
Stefanie Molin
}
