% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{time_order_for_plots}
\alias{time_order_for_plots}
\title{Time of Month Factor}
\usage{
time_order_for_plots(reverse = FALSE)
}
\arguments{
\item{reverse}{Boolean whether or not to reverse the order. Defaults to FALSE}
}
\value{
A vector
}
\description{
Factor of ordered time of month for graphing
}
\author{
Stefanie Molin
}
