% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_curves.R
\name{tpr_threshold}
\alias{tpr_threshold}
\title{Get True Positive Rate using Thresholds}
\usage{
tpr_threshold(data, threshold, prediction_column)
}
\arguments{
\item{data}{Dataframe with a prediction and is_alert column}

\item{threshold}{Threshold for the prediction to be an alert (greater than or equal to)}

\item{prediction_column}{Column name of the prediction values}
}
\value{
TPR
}
\description{
Calculate true positive rate (TPR)
}
\author{
Stefanie Molin
}
