% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrices.R
\name{true_negatives}
\alias{true_negatives}
\title{Get True Negatives}
\usage{
true_negatives(conf_mat)
}
\arguments{
\item{conf_mat}{A confusion matrix}
}
\value{
The number of true negatives
}
\description{
Calculate true negatives
}
\author{
Stefanie Molin
}
