% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_curves.R
\name{true_positive_rate}
\alias{true_positive_rate}
\title{Get True Positive Rate}
\usage{
true_positive_rate(true_positives, false_negatives)
}
\arguments{
\item{true_positives}{Number of true positives}

\item{false_negatives}{Number of false negatives}
}
\value{
TPR
}
\description{
Calculate true positive rate (TPR)
}
\author{
Stefanie Molin
}
