% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrices.R
\name{true_positives}
\alias{true_positives}
\title{Get True Positives}
\usage{
true_positives(conf_mat)
}
\arguments{
\item{conf_mat}{A confusion matrix}
}
\value{
The number of true positives
}
\description{
Calculate true positives
}
\author{
Stefanie Molin
}
