% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_study_results.R
\name{TS_alert_counts}
\alias{TS_alert_counts}
\title{TS Alert Counts}
\usage{
TS_alert_counts(beginning, middle, end)
}
\arguments{
\item{beginning}{Dataframe of TS results for beginning of the month}

\item{middle}{Dataframe of TS results for middle of the month}

\item{end}{Dataframe of TS results for end of the month}
}
\value{
A list containing dataframe of TS alert counts, a graph of the results, and a boxplot of the distribution
}
\description{
Calculate TS client alert counts for case study.
}
\author{
Stefanie Molin
}
