% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_study_results.R
\name{TS_alert_counts_by_metric}
\alias{TS_alert_counts_by_metric}
\title{TS Metric Alert Counts}
\usage{
TS_alert_counts_by_metric(beginning, middle, end)
}
\arguments{
\item{beginning}{Dataframe of TS results for beginning of the month}

\item{middle}{Dataframe of TS results for middle of the month}

\item{end}{Dataframe of TS results for end of the month}
}
\value{
A list containing dataframe of event name alert counts, a dataframe of site type alert counts, and a graph of the results
}
\description{
Calculate TS alert counts by metric for case study.
}
\author{
Stefanie Molin
}
