% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrices.R
\name{TS_metrics_confusion_matrix}
\alias{TS_metrics_confusion_matrix}
\title{TS Metrics Confusion Matrix}
\usage{
TS_metrics_confusion_matrix(beginning, middle, end, logs)
}
\arguments{
\item{beginning}{Dataframe of TS results for beginning of the month}

\item{middle}{Dataframe of TS results for middle of the month}

\item{end}{Dataframe of TS results for end of the month}

\item{logs}{Dataframe of logs from Metis}
}
\value{
A dataframe
}
\description{
Calculate TS confusion matrix by metric
}
\author{
Stefanie Molin
}
