% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{twitter_anomaly_detection_AS}
\alias{twitter_anomaly_detection_AS}
\title{AS Twitter Anomaly Detection}
\usage{
twitter_anomaly_detection_AS(data, period = NULL)
}
\arguments{
\item{data}{Dataframe in case study format}

\item{period}{Integer defining the length of the period for seasonal decomposition}
}
\value{
A dataframe
}
\description{
Test AS data for anomaly
}
\author{
Stefanie Molin
}
