% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{twitter_anomaly_detection_TS}
\alias{twitter_anomaly_detection_TS}
\title{TS Twitter Anomaly Detection}
\usage{
twitter_anomaly_detection_TS(data, period = NULL)
}
\arguments{
\item{data}{Dataframe in case study format}

\item{period}{Integer defining the length of the period for seasonal decomposition}
}
\value{
A dataframe
}
\description{
Check for TS anomalies using the Twitter algorithm
}
\author{
Stefanie Molin
}
