% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmelters.R
\name{unmelt_AS}
\alias{unmelt_AS}
\title{Unmelt AS Data Pulled from Metis}
\usage{
unmelt_AS(data)
}
\arguments{
\item{data}{Dataframe of AS data pulled from Metis}
}
\value{
Unmelted AS dataframe
}
\description{
Unmelts data pulled from Metis to use with other anomaly detection methods.
}
\note{
This function relies on the structure of the query that provides "data". 
If changing column names, adding or subtracting columns, be sure to update here as well.
}
\author{
Stefanie Molin
}
