% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmelters.R
\name{unmelt_TS}
\alias{unmelt_TS}
\title{Unmelt TS Data Pulled from Metis}
\usage{
unmelt_TS(data)
}
\arguments{
\item{data}{Dataframe of TS data pulled from Metis}
}
\value{
Unmelted TS dataframe
}
\description{
Unmelts data pulled from Metis to use with other anomaly detection methods.
}
\note{
This function relies on the structure of the query that provides "data". 
If changing column names, adding or subtracting columns, be sure to update here as well.
}
\author{
Stefanie Molin
}
