% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_library.R
\name{violin_plot_AS}
\alias{violin_plot_AS}
\title{Violin Plot for AS Data}
\usage{
violin_plot_AS(df)
}
\arguments{
\item{df}{Dataframe to build the plot with}
}
\value{
A ggplot object
}
\description{
Draw a violin plot for AS data
}
\author{
Stefanie Molin
}
