% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_library.R
\name{violin_plot_RexT}
\alias{violin_plot_RexT}
\title{Violin Plot for RexT Data}
\usage{
violin_plot_RexT(df)
}
\arguments{
\item{df}{Dataframe to build the plot with}
}
\value{
A ggplot object
}
\description{
Draw a violin plot for RexT data
}
\author{
Stefanie Molin
}
