% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_library.R
\name{violin_plot_TS}
\alias{violin_plot_TS}
\title{Violin Plot for TS Data}
\usage{
violin_plot_TS(df)
}
\arguments{
\item{df}{Dataframe to build the plot with}
}
\value{
A list of ggplot objects (one for event name plots and one for site type plots)
}
\description{
Draw Violin Plots for TS data
}
\author{
Stefanie Molin
}
