% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrices.R
\name{western_electric_AS_confusion_matrix}
\alias{western_electric_AS_confusion_matrix}
\title{AS Alerts Period Confusion Matrix for Western Electric Rules}
\usage{
western_electric_AS_confusion_matrix(beginning, middle, end, logs, rule,
  rules_order)
}
\arguments{
\item{beginning}{Dataframe of AS results for beginning of the month}

\item{middle}{Dataframe of AS results for middle of the month}

\item{end}{Dataframe of AS results for end of the month}

\item{logs}{Dataframe of logs from Metis}

\item{rule}{Name(s) of the rule(s) to check for. This must match the output of the check for alerts function, i.e. "Rule 1" or "Stratification rule" 
(the "violation" part will be added automatically)}

\item{rules_order}{Vector of rules from first checked to last}
}
\value{
A dataframe
}
\description{
Calculate AS confusion matrix by date
}
\author{
Stefanie Molin
}
