% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_study_results.R
\name{western_electric_AS_metric_counts}
\alias{western_electric_AS_metric_counts}
\title{Western Electric Rules AS Metric Alert Counts}
\usage{
western_electric_AS_metric_counts(beginning, middle, end, rule)
}
\arguments{
\item{beginning}{Dataframe of AS results for beginning of the month}

\item{middle}{Dataframe of AS results for middle of the month}

\item{end}{Dataframe of AS results for end of the month}

\item{rule}{Rule to filter for}
}
\value{
A list containing dataframe of client alert counts, a dataframe of campaign alert counts, and a graph of the results
}
\description{
Calculate AS alert counts by metric for case study using Western Electric Rules.
}
\author{
Stefanie Molin
}
