% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrices.R
\name{western_electric_exec_confusion_matrix}
\alias{western_electric_exec_confusion_matrix}
\title{Territory RexT Period Confusion Matrix for Western Electric Rules}
\usage{
western_electric_exec_confusion_matrix(data, logs, beginning_dates,
  middle_dates, end_dates, rule, rules_order)
}
\arguments{
\item{data}{Dataframe of territory RexT results}

\item{logs}{Dataframe of logs from Metis}

\item{beginning_dates}{Vector of dates for beginning of the month}

\item{middle_dates}{Vector of dates for middle of the month}

\item{end_dates}{Vector of dates for end of the month}

\item{rule}{Name(s) of the rule(s) to check for. This must match the output of the check for alerts function, i.e. "Rule 1" or "Stratification rule" 
(the "violation" part will be added automatically)}

\item{rules_order}{Vector of rules from first checked to last}
}
\value{
A dataframe
}
\description{
Calculate Territory RexT confusion matrix by date
}
\author{
Stefanie Molin
}
