% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/western_electric_rules.R
\name{western_electric_pipeline_AS}
\alias{western_electric_pipeline_AS}
\title{Run through AS Western Electricc Pipeline for Case Study}
\usage{
western_electric_pipeline_AS(data)
}
\arguments{
\item{data}{Dataframe of daily data to check for anomalies.}
}
\value{
A dataframe with the results
}
\description{
Clean data, impute the median for missing dates, and detect anomalies for all client/campaign/kpi combinations.
}
\note{
This function relies on the structure of the query that provides "data". 
If changing column names, adding or subtracting columns, be sure to update here as well.

Client/campaign must have TAC in each of the last 2 days and 25 rows in the database in the last 30 days
}
\author{
Stefanie Molin
}
