% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/western_electric_rules.R
\name{western_electric_pipeline_RexT}
\alias{western_electric_pipeline_RexT}
\title{RexT Western Electric Rules Anomaly Detection}
\usage{
western_electric_pipeline_RexT(data)
}
\arguments{
\item{data}{Dataframe in case study format}
}
\value{
A dataframe
}
\description{
Check for Territory RexT anomalies using the Western Electric rules
}
\author{
Stefanie Molin
}
