% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/western_electric_rules.R
\name{western_electric_pipeline_TS}
\alias{western_electric_pipeline_TS}
\title{Western Electric Rules TS Alerts}
\usage{
western_electric_pipeline_TS(data)
}
\arguments{
\item{data}{Dataframe in case study format}
}
\value{
A dataframe
}
\description{
Check for TS anomalies using the Western Electric rules
}
\author{
Stefanie Molin
}
