% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrices.R
\name{western_electric_rule_confusion_matrix}
\alias{western_electric_rule_confusion_matrix}
\title{Confusion Matrix by Rule for Western Electric Rules}
\usage{
western_electric_rule_confusion_matrix(AS_data, TS_data, RexT_data, logs, rule,
  rules_order)
}
\arguments{
\item{AS_data}{Dataframe of AS results for beginning of the month}

\item{TS_data}{Dataframe of AS results for middle of the month}

\item{RexT_data}{Dataframe of AS results for end of the month}

\item{logs}{Dataframe of logs from Metis}

\item{rule}{Name(s) of the rule(s) to check for. This must match the output of the check for alerts function, i.e. "Rule 1" or "Stratification rule" 
(the "violation" part will be added automatically)}

\item{rules_order}{Vector of rules from first checked to last}
}
\value{
A dataframe
}
\description{
Calculate confusion matrix for given rule across AS, TS, and RexT
}
\author{
Stefanie Molin
}
