% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/western_electric_rules.R
\name{western_electric_rules}
\alias{western_electric_rules}
\title{Western Electric Rules}
\usage{
western_electric_rules(data)
}
\arguments{
\item{data}{Vector to test}
}
\value{
A list with "anomaly"--a boolean indicating whether or not yesterday was an alert--and
"reason"--name of the rule that was violated if there was a violation, otherwise "N/A"
}
\description{
Determine if yesterday is an anomaly according to the Western Electric Rules.
}
\note{
More info here: \url{https://en.wikipedia.org/wiki/Western_Electric_rules} and here: 
\url{https://en.wikipedia.org/wiki/Nelson_rules}
}
\author{
Stefanie Molin
}
