% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_study_results.R
\name{western_electric_TS_metric_counts}
\alias{western_electric_TS_metric_counts}
\title{Western Electric Rules TS Metric Alert Counts}
\usage{
western_electric_TS_metric_counts(beginning, middle, end, rule)
}
\arguments{
\item{beginning}{Dataframe of TS results for beginning of the month}

\item{middle}{Dataframe of TS results for middle of the month}

\item{end}{Dataframe of TS results for end of the month}

\item{rule}{Rule to filter for}
}
\value{
A list containing dataframe of event name alert counts, a dataframe of site type alert counts, and a graph of the results
}
\description{
Calculate TS alert counts by metric for case study using Western Electric Rules.
}
\author{
Stefanie Molin
}
